backend/pyproject.toml
[project]
name = "pickleballsinglessocialv2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.6.2",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.3",
    "stripe>=14.4.1",
    "whitenoise>=6.12.0",
]
